Added tests for the BoardContainer.
I think there are ways to test this without making the map functions public,
but last time I looked into this I had the feeling the "react-redux" lib
has not been written with testign in mind.

Its a bit like a composition root DI framework,
where we wire up implmentations to interfaces, but the implementation is defined
inline, and cant really be tested in isolation.
So making them public and verifying their behaviour in a small pice
like this feels like a good compromise